feat: cloning entity
